# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI TimeShards API",
    "version": "0.1.0",
    "description": "REST API for time tracking and access control. Narrative reference: docs/API.md. Machine-readable spec served at GET /api/v1/openapi.json."
  },
  "servers": [{ "url": "http://127.0.0.1:47821", "description": "Default local bind" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token from POST /api/v1/auth/login"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/time/work-calendars/{id}/days": {
      "get": {
        "summary": "Calendar days in range",
        "parameters": [
          { "name": "from", "in": "query", "required": true, "schema": { "type": "string", "format": "date" } },
          { "name": "to", "in": "query", "required": true, "schema": { "type": "string", "format": "date" } }
        ]
      }
    },
...
  },
  "security": [{ "bearerAuth": [] }]
}